# packages/openapi-generator/tmp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lonestone API",
    "description": "The Lonestone API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "paths": {
...
    "/api/groups": {
      "get": {
        "operationId": "CommunityController_getGroups",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sector",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/groups/{id}/join": {
      "post": {
        "operationId": "CommunityController_joinGroup",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/groups/{id}/leave": {
      "delete": {
        "operationId": "CommunityController_leaveGroup",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/groups/{id}/publications": {
      "get": {
        "operationId": "CommunityController_getPublications",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "CommunityController_createPublication",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "PRODUCT_ANNOUNCEMENT",
                      "TECHNICAL_QUESTION",
                      "MARKET_ALERT",
                      "TRAINING_SHARE"
                    ]
                  },
                  "content": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  },
                  "mediaUrls": {
                    "maxItems": 4,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                },
                "required": [
                  "type",
                  "content"
                ],
                "additionalProperties": false,
                "title": "CreatePublication"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/groups/publications/{id}/report": {
      "post": {
        "operationId": "CommunityController_reportPublication",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/groups/publications/{id}/share-url": {
      "get": {
        "operationId": "CommunityController_getShareUrl",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}